+++ src/service/TRex/tests/netstandard/VSS.TRex.Tests/IO/SlabAllocatedArrayPoolTests.cs
@@ -47,7 +47,7 @@ namespace VSS.TRex.Tests.IO
         rental.Elements.Should().NotBeNull();
         rental.Offset.Should().BeGreaterOrEqualTo(0);
         rental.OffsetPlusCount.Should().BeGreaterOrEqualTo(0);
-        rental.SlabIndex.Should().Be((byte)(i == 0 ? TRexSpan<CellPass>.NO_SLAB_INDEX : 0));
+        rental.SlabIndex.Should().Be((ushort)(i == 0 ? TRexSpan<CellPass>.NO_SLAB_INDEX : 0));
 
         pool.Return(rental); // Release rental so as not to pollute expected pool allocated status
       }
@@ -84,7 +84,7 @@ namespace VSS.TRex.Tests.IO
       for (int i = 0; i < DEFAULT_TEST_SLAB_ALLOCATED_POOL_SIZE; i++)
       {
         var rental = pool.Rent(i);
-        rental.SlabIndex.Should().Be((byte)(i == 0 ? TRexSpan<CellPass>.NO_SLAB_INDEX : 0)); 
+        rental.SlabIndex.Should().Be((ushort)(i == 0 ? TRexSpan<CellPass>.NO_SLAB_INDEX : 0)); 
 
         pool.Return(rental);
         rental.MarkReturned();
